Migrates alert and keyboard to json_wire_spec.
