+++ lib/src/sync/json_wire_spec/keyboard.dart
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import 'common.dart';
-import 'web_driver.dart';
+import '../common.dart';
+import '../web_driver.dart';
 
 class Keyboard {
   static const String nullChar = '\uE000';
